Use float property type

+++ examples/epen_graph_editor/include/data_models/Scalar_Float_Base.hpp
@@ -4,9 +4,9 @@ class Scalar_Float_Base : public UIBufferBase
 {
     Q_OBJECT
 public:
-    Q_PROPERTY(double Default_Value MEMBER _defaultValue NOTIFY propertyChanged)
-    Q_PROPERTY(double Min MEMBER _min NOTIFY propertyChanged)
-    Q_PROPERTY(double Max MEMBER _max NOTIFY propertyChanged)
+    Q_PROPERTY(float Default_Value MEMBER _defaultValue NOTIFY propertyChanged)
+    Q_PROPERTY(float Min MEMBER _min NOTIFY propertyChanged)
+    Q_PROPERTY(float Max MEMBER _max NOTIFY propertyChanged)
 
 private:
     float _defaultValue;
