+++ examples/epen_graph_editor/include/data_models/Scalar_Float_Fixed.hpp
@@ -4,7 +4,7 @@ class Scalar_Float_Fixed : public UIBufferBase
 {
     Q_OBJECT
 public:
-    Q_PROPERTY(double Value MEMBER _value NOTIFY propertyChanged)
+    Q_PROPERTY(float Value MEMBER _value NOTIFY propertyChanged)
     QString caption() const override { return QStringLiteral("float Fixed Buffer"); }
 
 private:
